+++ Generals/Code/GameEngine/Source/Common/Thing/ThingTemplate.cpp
@@ -1114,7 +1114,7 @@ ThingTemplate::~ThingTemplate()
 //=============================================================================
 void ThingTemplate::resolveNames()
 {
-	Int i, j;
+	size_t i, j;
 
 	for (i = 0; i < m_prereqInfo.size(); i++)
 	{
