+++ Generals/Code/GameEngineDevice/Source/W3DDevice/GameClient/W3DDisplay.cpp
@@ -1828,7 +1828,7 @@ AGAIN:
 					Int height = TheDisplay->getHeight() * .9;
 
 					Int width;
-					if( displayString->getWidth() > TheDisplay->getWidth() )
+					if( displayString->getWidth() > static_cast<Int>(TheDisplay->getWidth()) )
 						width = 20;
 					else
 						width = ( TheDisplay->getWidth() - displayString->getWidth() ) / 2;
